exclude t/ci for format checking
As we discussed in the call on March 29, adding
a tag in the checkpatch.sh to ignore files in the
t/ci directory.  The checkpatch is basically pointed
at C source code.  There is no C source in that
directory, only bash scripts, and the 80-character
line limit is not reasonable for bash scripts that
need to construct long command lines.